requirements.txt
bingbong
sseclient-py
chromadb
pydantic-settings
google-generativeai
pypdf2
PyMuPDF
gradio
requests
toml
datasets
flatdict
APScheduler
